requirements.txt
# Python依赖包
python-docx==0.8.11
fastapi==0.95.0
uvicorn==0.22.0
pyyaml==6.0.1
pydantic>=2.0.0
